AAIDC_Project_1/Submission/requirements.txt
chromadb==0.4.22
sentence-transformers>=2.3.0
python-dotenv==1.0.0
langchain-core==0.1.23
langchain-openai==0.0.5
langchain-groq==0.0.1
langchain-google-genai==0.0.5
langchain-community==0.0.20
